Fix bug in deleteForwardLine

+++ packages/slate-delete-one-line/src/index.js
@@ -2,16 +2,31 @@
 import HotKeys from 'slate-hotkeys';
 import { type Change } from 'slate';
 import deleteLineBackward from './changes/delete-line-backward';
+import deleteLineForward from './changes/delete-line-forward';
 
-function onKeyDown(event: SyntheticKeyboardEvent<*>, change: Change) {
-    const { value } = change;
-    if (value.isExpanded) return undefined;
-    if (HotKeys.isDeleteLineBackward(event)) {
-        return deleteLineBackward(change);
-    }
-    return undefined;
+type Options = {
+    backward?: *,
+    forward?: *
+};
+
+type HandleKeyDown = (SyntheticKeyboardEvent<*>, Change) => *;
+
+function onKeyDown(options: Options): HandleKeyDown {
+    const { forward, backward } = options;
+    return (event: SyntheticKeyboardEvent<*>, change: Change) => {
+        const { value } = change;
+        if (value.isExpanded) return undefined;
+        if (HotKeys.isDeleteLineBackward(event) && backward) {
+            return deleteLineBackward(change);
+        }
+        if (HotKeys.isDeleteLineForward(event) && forward) {
+            return deleteLineForward(change);
+        }
+        return undefined;
+    };
 }
 
-export default function() {
-    return { onKeyDown };
+export default function(options?: Options = {}) {
+    const { backward = true, forward = true } = options;
+    return { onKeyDown: onKeyDown({ backward, forward }) };
 }
